Refactor NodeEditor event bus implementation and move to separate file
